Fix auto-tuner
